fix: move some templates to files
Some of the templates cannot be embedded because they themselves contain jinja. For now this
is a quick hack to make them work. If we need any in the long term i will clean this up and
find a less hacky solution, there is quite the chance that we will not need the currently
affected tempaltes anyway in the long run